done ctsp admin

+++ BanDoDienTu_Nhom06_N03/BanDoDienTu_Nhom06_N03/Areas/Admin/Controllers/ProductDetailsController.cs
@@ -6,6 +6,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using BanDoDienTu_Nhom06_N03.Models;
+using AspNetCoreHero.ToastNotification.Abstractions;
+using X.PagedList;
 
 namespace BanDoDienTu_Nhom06_N03.Areas.Admin.Controllers
 {
@@ -14,17 +16,22 @@ namespace BanDoDienTu_Nhom06_N03.Areas.Admin.Controllers
     {
         private readonly BanDoDienTuContext _context;
 
-        public ProductDetailsController(BanDoDienTuContext context)
+        public INotyfService _notyfService { get; }
+
+        public ProductDetailsController(BanDoDienTuContext context, INotyfService notyfService)
         {
             _context = context;
+            _notyfService = notyfService;
         }
 
         // GET: Admin/ProductDetails
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(int? page)
         {
-              return _context.ChiTietSps != null ? 
-                          View(await _context.ChiTietSps.ToListAsync()) :
-                          Problem("Entity set 'BanDoDienTuContext.ChiTietSps'  is null.");
+            int pageSize = 10;
+            int pageNumber = page == null || page < 0 ? 1 : page.Value;
+            var listProductDetails = _context.ChiTietSps.ToList();
+            PagedList<ChiTietSp> res = new PagedList<ChiTietSp>(listProductDetails, pageNumber, pageSize);
+            return View(res);
         }
 
         // GET: Admin/ProductDetails/Details/5
@@ -60,9 +67,17 @@ namespace BanDoDienTu_Nhom06_N03.Areas.Admin.Controllers
         {
             if (ModelState.IsValid)
             {
-                _context.Add(chiTietSp);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if(_context.ChiTietSps.FirstOrDefault(x => x.MaSp == chiTietSp.MaSp)== null)
+                {
+                    _context.Add(chiTietSp);
+                    await _context.SaveChangesAsync();
+                    _notyfService.Success("Thêm thành công");
+                    return RedirectToAction(nameof(Index));
+                }
+                else
+                {
+                    _notyfService.Success("Mã sản phẩm đã tồn tại");
+                }
             }
             return View(chiTietSp);
         }
@@ -113,8 +128,10 @@ namespace BanDoDienTu_Nhom06_N03.Areas.Admin.Controllers
                         throw;
                     }
                 }
+                _notyfService.Success("Sửa thành công");
                 return RedirectToAction(nameof(Index));
             }
+            _notyfService.Error("Sửa không thành công");
             return View(chiTietSp);
         }
 
@@ -145,14 +162,21 @@ namespace BanDoDienTu_Nhom06_N03.Areas.Admin.Controllers
             {
                 return Problem("Entity set 'BanDoDienTuContext.ChiTietSps'  is null.");
             }
-            var chiTietSp = await _context.ChiTietSps.FindAsync(id);
-            if (chiTietSp != null)
+            var sp = _context.SanPhams.FirstOrDefault(x => x.MaSp== id);
+            if(sp == null)
             {
-                _context.ChiTietSps.Remove(chiTietSp);
+                var chiTietSp = await _context.ChiTietSps.FindAsync(id);
+                if (chiTietSp != null)
+                {
+                    _context.ChiTietSps.Remove(chiTietSp);
+                }
+
+                await _context.SaveChangesAsync();
+                _notyfService.Success("Xóa thành công");
+                return RedirectToAction(nameof(Index));
             }
-            
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            _notyfService.Error("Xóa sản phẩm có mã" + id + " trước khi xóa chi tiết");
+            return RedirectToAction(nameof(Delete));
         }
 
         private bool ChiTietSpExists(string id)
